requirements3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements3.txt requirements3.in
#

click==6.6                # via pip-tools
first==2.0.1              # via pip-tools
invoke==0.12.2
jinja2==2.8
MarkupSafe==0.23          # via jinja2
nose==1.3.7
PasteDeploy==1.5.2        # via pyramid
pip-tools==1.7.0
pyramid==1.7
pyyaml==3.11
repoze.lru==0.6           # via pyramid
requests==2.10.0
six==1.10.0               # via pip-tools
translationstring==1.3    # via pyramid
venusian==1.0             # via pyramid
waitress==0.9.0
WebOb==1.6.1              # via pyramid
zope.deprecation==4.1.2   # via pyramid
zope.interface==4.2.0     # via pyramid

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via pyramid, zope.deprecation, zope.interface
